pyproject.toml
[tool.poetry]
name = "StockAnalysis"
version = "0.1.0"
description = ""
authors = ["ccortezaguilera <10718803+ccortezaguilera@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
asyncio = "^3.4.3"
yfinance = "^0.1.54"
fastapi = "^0.61.1"
uvicorn = "^0.12.1"
ujson = "^4.0.1"
Jinja2 = "^2.11.2"
typing-extensions = "^3.7.4"
aiofiles = "^0.5.0"
pydantic = {version = "^1.6.1", extras = ["dotenv"]}
SQLAlchemy = "^1.3.20"
alembic = "^1.4.3"
pyhumps = "^1.6.1"

[tool.poetry.dev-dependencies]
black = "^21.4b0"
pytest = "^5.2"
commitizen = "^2.4.1"
pre-commit = "^2.7.1"
PyYAML = "^5.3.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
